Moving responsibilities to individual sbt classes instead of in the sbt bom tool.
