+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/bomtool/sbt/SbtReportParser.groovy
@@ -1,0 +1,84 @@
+/*
+ * Copyright (C) 2017 Black Duck Software Inc.
+ * http://www.blackducksoftware.com/
+ * All rights reserved.
+ *
+ * This software is the confidential and proprietary information of
+ * Black Duck Software ("Confidential Information"). You shall not
+ * disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Black Duck Software.
+ */
+package com.blackducksoftware.integration.hub.detect.bomtool.sbt;
+
+import groovy.util.slurpersupport.GPathResult;
+
+public class SbtReport {
+    String organisation;
+
+    String module;
+
+    String revision;
+
+    String configuration;
+
+    List<SbtModule> dependencies;
+}
+
+public class SbtModule {
+    String organisation;
+
+    String name;
+
+    List<SbtRevision> revisions;
+}
+
+public class SbtRevision {
+    String name;
+
+    List<SbtCaller> callers;
+}
+
+public class SbtCaller {
+    String callerOrganisation;
+
+    String callerName;
+
+    String callerRevision;
+}
+
+public class SbtReportParser {
+    public SbtReport parseReportFromXml(GPathResult xmlReport) {
+        SbtReport report = new SbtReport();
+        report.organisation = xmlReport.info.@organisation.toString();
+        report.module = xmlReport.info.@module.toString();
+        report.revision = xmlReport.info.@revision.toString();
+        report.configuration = xmlReport.info.@conf.toString();
+        report.dependencies = new ArrayList<SbtModule>();
+
+        xmlReport.dependencies.module.each{xmlModule ->
+            SbtModule module = new SbtModule();
+            module.name = xmlModule.@name.toString()
+            module.organisation = xmlModule.@organisation.toString()
+            module.revisions = new ArrayList<SbtRevision>();
+            report.dependencies.add(module);
+
+            xmlModule.revision.each{  xmlRevision ->
+                SbtRevision revision = new SbtRevision();
+                revision.name = xmlRevision.@name.toString()
+                revision.callers = new ArrayList<SbtCaller>();
+                module.revisions.add(revision);
+
+                xmlRevision.caller.each { xmlCaller ->
+                    SbtCaller caller = new SbtCaller();
+                    caller.callerOrganisation = xmlCaller.@organisation.toString()
+                    caller.callerName = xmlCaller.@name.toString()
+                    caller.callerRevision = xmlCaller.@callerrev.toString()
+                    revision.callers.add(caller)
+                }
+            }
+        }
+
+        report
+    }
+}
